% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author_caption.R
\name{add_author_attribution}
\alias{add_author_attribution}
\title{Add author caption to graphic}
\usage{
add_author_attribution(p = NULL, author = "Josh Persi", family = NULL)
}
\arguments{
\item{p}{A ggplot object.}

\item{author}{A character string specifying the author's name. Default is "Josh Persi".}

\item{family}{A character string specifying the font family for the caption text. If NULL, uses the font family from the plot theme.}
}
\value{
NULL. The function draws the caption on the current graphics device.
}
\description{
Add the name of the author to the bottom right corner of a ggplot graphic.
}
\examples{
temp_file <- fs::file_temp(ext = ".png")

ragg::agg_png(temp_file, width = 6, height = 6, units = "in", res = 300)

p <- ggplot2::ggplot(penguins, ggplot2::aes(bill_dep, bill_len)) +
  ggplot2::geom_point() +
  ggplot2::theme(
    plot.margin = ggplot2::margin(0.25, 0.25, 0.375, 0.25, unit = "in")
)

base::print(p)

add_author_attribution(p)

grDevices::dev.off()

}
