% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_source_caption.R
\name{add_source_attribution}
\alias{add_source_attribution}
\title{Add data source caption to graphic}
\usage{
add_source_attribution(p = NULL, source, family = NULL)
}
\arguments{
\item{p}{A ggplot object.}

\item{family}{A character string specifying the font family for the caption text. If NULL, uses the font family from the plot theme.}
}
\value{
NULL. The function draws the caption on the current graphics device.
}
\description{
Add the name of the data source to the bottom left corner of a ggplot graphic.
}
\examples{
temp_file <- fs::file_temp(ext = ".png")

ragg::agg_png(temp_file, width = 6, height = 6, units = "in", res = 300)

p <- ggplot2::ggplot(penguins, ggplot2::aes(bill_dep, bill_len)) +
  ggplot2::geom_point() +
  ggplot2::theme(
    plot.margin = ggplot2::margin(0.25, 0.25, 0.375, 0.25, unit = "in")
  )

base::print(p)

add_source_attribution(p, "An example source")

grDevices::dev.off()

}
