% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{ansi_to_html}
\alias{ansi_to_html}
\title{Convert ANSI color codes to HTML tags.}
\usage{
ansi_to_html(x)
}
\arguments{
\item{x}{Text snippet.}
}
\value{
Text snippet with ANSI color codes replaced by HTML tags.
}
\description{
Convert ANSI color codes to HTML tags.
}
