% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cpaste}
\alias{cpaste}
\alias{cat0}
\alias{ccat}
\alias{ccat0}
\alias{cinform}
\alias{cinform0}
\title{Colored cat.}
\usage{
cpaste(..., sep = " ", default = "default")

cat0(..., sep = "")

ccat(..., sep = " ", default = "default")

ccat0(..., default = "default")

cinform(..., sep = " ", default = "default")

cinform0(..., default = "default")
}
\description{
Colored cat.
}
