% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{colorize}
\alias{colorize}
\alias{colorize_}
\title{Add color to text.}
\usage{
colorize(text, ..., default = "default")

colorize_(text, repl, default = "default")
}
\arguments{
\item{text}{character string to amend.}

\item{...}{substrings with colors assigned as names.}

\item{default}{default color.}

\item{repl}{named vector or list; names are colors, values are substrings.}
}
\value{
\code{text} with ANSI control sequences around colorized substrings.
}
\description{
Transforms the input \code{text} by adding colors to substrings specified
in \code{...}.
}
\examples{
colorize("text with words", yellow = "with")
colorize("longer text with many words", yellow = "text", red = "many")
}
