% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{colourize}
\alias{colourize}
\alias{colorize}
\title{Add color to text.}
\usage{
colourize(text, ..., default = "default")

colorize(text, ..., default = "default")
}
\arguments{
\item{text}{character string to amend.}

\item{...}{substrings with colors assigned as names.}

\item{default}{default color.}
}
\value{
\code{text} with ANSI control sequences around colorized substrings.
}
\description{
Transforms the input \code{text} by adding colors to substrings specified
in \code{...}.
}
\examples{
colourize("text with words", yellow = "with")
colourize("longer text with many words", yellow = "text", red = "many")
}
