% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{compareQueryWithSubjectSequences}
\alias{compareQueryWithSubjectSequences}
\title{Compare query to subject sequences}
\usage{
compareQueryWithSubjectSequences(querySeqs, subjectSeqs)
}
\arguments{
\item{character}{vector of querySeqs}

\item{character}{vector of subjectSeqs}
}
\value{
data.frame with the closest match to each sequence
}
\description{
This method compares query sequences to subject sequences and returns a data.frane with the closest match as rows of subject sequences
}

