% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{compareRNASeqInChip}
\alias{compareRNASeqInChip}
\title{Compare sequences to chip oligos}
\usage{
compareRNASeqInChip(rnaSeqCandidateSequences, type = "global-local", ...)
}
\arguments{
\item{...}{further arguments to pairwiseAlignment method}

\item{character}{vector of sequences to be analysed}

\item{character}{vector of mapping type for pairwiseAlignment method (default: global-local means global oligo and local in the RNA-Seq)}
}
\value{
data.frame with the closest match to each sequence
}
\description{
compares sequences to the oligonucleotide sequence (reverseComplement of oligo)
}

