% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\docType{data}
\name{configs}
\alias{configs}
\alias{plots}
\alias{estimates}
\alias{trends}
\alias{agegroups}
\title{Load config files}
\format{An object of class \code{list} of length 4.}
\usage{
plots

estimates

trends

agegroups
}
\description{
These variables store the data found in the config files. This config data
helps build the server and ui of the Tabby application
}
\keyword{datasets}
