% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{new_vector}
\alias{new_vector}
\alias{new_map}
\alias{new_set}
\title{STL-like containers.}
\usage{
new_vector(..., data = list())

new_map(..., data = list())

new_set(..., data = list())
}
\arguments{
\item{...}{Elements for the container, named for \code{map}.}

\item{data}{Alternative way to specify elements.}
}
\description{
STL-like containers.
}
