% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\alias{print.counter}
\title{State-full counter object.}
\usage{
counter(start = 0)

\method{print}{counter}(x, ...)
}
\arguments{
\item{start}{initial value}

\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A counter function, each call on which returns a subsequent
number.
}
\description{
State-full counter object.
}
\examples{
c <- counter()
c() # returns 0
c() # returns 1
c() # returns 2
}
