% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cpaste}
\alias{cpaste}
\title{Colored cat.}
\usage{
cpaste(..., sep = " ", default = "default")
}
\arguments{
\item{...}{Strings to be printed; optionally named with color names.}

\item{sep}{Separator.}

\item{default}{Default color; \code{"default"} means no color.}
}
\description{
Colored cat.
}
\seealso{
colourize
}
