% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{createDoubleVenn}
\alias{createDoubleVenn}
\title{Wrapper for draw.pairwise.venn function in VennDiagram}
\usage{
createDoubleVenn(idsA, idsB, idsC, category = c("First", "Second"),
  fill = c("blue", "red"), cat.col = c("blue", "red"), lty = "blank",
  cex = 2, cat.cex = 2, ...)
}
\arguments{
\item{category}{(see draw.pairwise.venn)}

\item{fill}{(see draw.pairwise.venn)}

\item{cat.col}{(see draw.pairwise.venn)}

\item{lty}{(see draw.pairwise.venn)}

\item{cex}{(see draw.pairwise.venn)}

\item{cat.cex}{(see draw.pairwise.venn)}

\item{ids}{of sample A}

\item{ids}{of sample B}
}
\value{
result object of draw.pairwise.venn
}
\description{
This function is a wrapper for draw.pairwise.venn.
}

