% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{createQuadVenn}
\alias{createQuadVenn}
\title{Wrapper for draw.quad.venn function in VennDiagram}
\usage{
createQuadVenn(idsA, idsB, idsC, idsD, category = c("First", "Second",
  "Third", "Fourth"), fill = c("orange", "red", "green", "blue"),
  cat.col = c("orange", "red", "green", "blue"), lty = "blank", cex = 2,
  cat.cex = 2, ...)
}
\arguments{
\item{category}{(see draw.quad.venn)}

\item{fill}{(see draw.quad.venn)}

\item{cat.col}{(see draw.quad.venn)}

\item{lty}{(see draw.quad.venn)}

\item{cex}{(see draw.quad.venn)}

\item{cat.cex}{(see draw.quad.venn)}

\item{ids}{of sample A}

\item{ids}{of sample B}

\item{ids}{of sample C}

\item{ids}{of sample D}
}
\value{
result object of (draw.quad.venn)
}
\description{
This function is a wrapper for draw.quad.venn
}

