% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{createTripleVenn}
\alias{createTripleVenn}
\title{Wrapper for draw.triple.venn function in VennDiagram}
\usage{
createTripleVenn(idsA, idsB, idsC, category = c("First", "Second", "Third"),
  fill = c("blue", "red", "green"), cat.col = c("blue", "red", "green"),
  lty = "blank", cex = 2, cat.cex = 2, ...)
}
\arguments{
\item{category}{(see draw.triple.venn)}

\item{fill}{(see draw.triple.venn)}

\item{cat.col}{(see draw.triple.venn)}

\item{lty}{(see draw.triple.venn)}

\item{cex}{(see draw.triple.venn)}

\item{cat.cex}{(see draw.triple.venn)}

\item{ids}{of sample A}

\item{ids}{of sample B}

\item{ids}{of sample C}
}
\value{
result object of (draw.triple.venn)
}
\description{
This function is a wrapper for draw.triple.venn
}

