% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.str.R
\name{datasets.str}
\alias{datasets.str}
\title{Structure of Available Datasets}
\usage{
datasets.str(package = NULL)
}
\arguments{
\item{package}{The package/packages containing the datasets of interest}
}
\value{
A data frame listing available data sets, invisibly
}
\description{
\code{datasets.str} returns the structure of available datasets
}
\details{
Datasets are often available in packages loaded into \code{R} and it is useful to know the structure of these datasets.  This function
shows the user the strucure of all available datasets in a specified package or packages.  (If the user does not specify a \code{package})
then the function searches over all available packages.
}
\examples{
datasets.str("datasets")
}
