% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{dbg}
\alias{dbg}
\alias{guard}
\title{Debug utilities.}
\usage{
dbg(...)

guard(fname = NULL)
}
\arguments{
\item{...}{Parts of the message, concatenated with an empty string.}

\item{fname}{Optional name of the function \code{guard()} is called from.}
}
\description{
Debug utilities.
}
