% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_defaults}
\alias{set_defaults}
\alias{get_default}
\title{Get/set default values as attributes.}
\usage{
set_defaults(x, ...)

get_default(x, name, default = NULL)
}
\arguments{
\item{x}{Object to get/set attributes on/from.}

\item{...}{Named attribute values.}

\item{name}{Name of the default value.}

\item{default}{Return if attribute not set on \code{x}.}
}
\description{
Get/set default values as attributes.
}
