% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{df_col_has_value}
\alias{df_col_has_value}
\title{check that dataframe column has data of specified type}
\usage{
df_col_has_value(df, column, expected_type)
}
\arguments{
\item{df}{data.frame}

\item{column}{column name to inspect}

\item{expected_type}{expected class of data in column}
}
\description{
check that dataframe column has data of specified type
}
