% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{df_to_list}
\alias{df_to_list}
\title{split dataframe row wise for iteration}
\usage{
df_to_list(df)
}
\arguments{
\item{df}{data.frame or tibble}
}
\value{
list
}
\description{
split dataframe row wise for iteration
}
