% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{downloadImportMiRNAAnnotation_mirBase}
\alias{downloadImportMiRNAAnnotation_mirBase}
\title{Download miRNA annotations from mirbase}
\usage{
downloadImportMiRNAAnnotation_mirBase(speciesFilter = "")
}
\arguments{
\item{species}{optional species identifier to subset the data (mmu for example)}
}
\value{
data.frame containing Name, Id, Name_wo_Species, Species
}
\description{
downloads the latest mirbase miRNA fasta file and extracts the names and identifier as data.frame
}

