% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{downloadImportMiRNAFamilies_mirBase}
\alias{downloadImportMiRNAFamilies_mirBase}
\title{Download and import MiRNAFamilies from mirbase}
\usage{
downloadImportMiRNAFamilies_mirBase()
}
\value{
list named list containing miRNAfam ID and miRNAfam name as name and a data.frame for each familiy containing miRNA name and ID
}
\description{
downloads the latest mirbase miRNA family file and imports it into R as list
}

