% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cutoff.R
\name{find_cutoff}
\alias{find_cutoff}
\title{Finds cells which are positive inputted signal
Make mode tree to calculate peaks
Find best bandwidth and get peak values for first two peaks, assuming data is bimodal
Find minimum between peaks}
\usage{
find_cutoff(
  data,
  n_peaks = 2,
  title = "",
  display = TRUE,
  approach = 1,
  threshold = NULL
)
}
\arguments{
\item{data}{Dataframe with 1 column containing dex calls and rownames as cell barcodes}

\item{n_peaks}{Number of peaks expected in data, Default 2 (Bimodal)}

\item{display}{Prints plot of cutoff}

\item{approach}{Which approach to use forcalculating threshold}

\item{threhold}{Manual numericl value, will bypass other approaches}
}
\description{
Finds cells which are positive inputted signal
Make mode tree to calculate peaks
Find best bandwidth and get peak values for first two peaks, assuming data is bimodal
Find minimum between peaks
}
