% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getEnrichedGenesFromDataset}
\alias{getEnrichedGenesFromDataset}
\title{Get Genes from GO-IDs}
\usage{
getEnrichedGenesFromDataset(geneDF, goIDs,
  geneDF_ensembleGeneColumnName = "UID", dataset = "mmusculus_gene_ensembl",
  annotDBIdb = "org.Mm.eg.db")
}
\arguments{
\item{geneDF}{dataframe containing the input genes (ensembl_gene_ids)}

\item{goIDs}{vector of GO-IDs}

\item{geneDF_ensembleGeneColumnName}{(The column name in the geneDF that represents the ensembl_gene_ids for merging)}

\item{dataset}{for biomart (human/mouse etc.) - default: mmusculus_gene_ensembl}

\item{annotDBIdb}{databse default: org.Mm.eg.db}
}
\value{
List of Data frames (subset of geneDF plus GOID) splitted by GO-ID
}
\description{
This function returns the genes present in the given dataset for each GO-ID
}

