% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getMiRNAClusterL}
\alias{getMiRNAClusterL}
\title{Prepares miRNA cluster}
\usage{
getMiRNAClusterL(species = "human")
}
\arguments{
\item{character}{default human, other possibility: mouse}
}
\value{
list that contains some annotation for each cluster and a named list (names miRNA cluster ID) and miRNA names as vector
}
\description{
loads clusters of miRNAs according to Wen-Ching Chan et al 2012
}

