% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-pipe_functions.R
\name{\%l0>\%}
\alias{\%l0>\%}
\title{\%l0>\%}
\usage{
lhs \%l0>\% rhs
}
\arguments{
\item{lhs}{arg to check}

\item{rhs}{arg to return if lhs length is 0}
}
\description{
If \code{length(lhs) == 0}, then change to \code{rhs}
}
\examples{
list(1, integer(0), character(0), numeric(0), NA,24) \%l0>\% 666
}
