% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Hist_Fn}
\alias{Hist_Fn}
\title{Copy of Hist_Fn}
\usage{
Hist_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?plot_histogram} to see list of arguments and usage
}
