% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{nth}
\alias{nth}
\alias{last}
\alias{first}
\alias{second}
\title{Indexing.}
\usage{
nth(x, n)

last(x)

first(x)

second(x)
}
\arguments{
\item{x}{Object to be indexed.}

\item{n}{Index value.}
}
\description{
Indexing.
}
