% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-is_missing.R
\name{is_missing}
\alias{is_missing}
\title{is_missing}
\usage{
is_missing(vec, check_if_finite = TRUE)
}
\arguments{
\item{vec}{vector}

\item{check_if_finite}{logical; If TRUE, then `Inf` tis considered as missing too}
}
\value{
logical vector
}
\description{
Sprawdzenie, czy elementy wektora nie są puste (NA, NULL, NaN, "" lub " "; opcjonalnie: +/-Inf)
checks if any vector element is empty (NA, NULL, NaN, "" or " "; optional: +/-Inf) and returns logical vector
}
\examples{
is_missing(vec = c(1,3, NA, NaN, Inf), check_if_finite = FALSE)
}
