% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{KDE}
\alias{KDE}
\title{Kernel Density Estimator}
\usage{
\code{KDE}
}
\arguments{
\item{data}{Input data for the kernel density estimator (a numeric vector)}

\item{weights}{Weights for the kernel density estimator (a numeric vector with the same length as the data)}

\item{bandwidth}{Bandwidth for the KDE; if \code{NULL} it is estimated using LOO-MLE}

\item{df}{Degrees-of-freedom for the T-distribution; if \code{NULL} it is estimated using LOO-MLE}

\item{df.norm}{Maximum recognised value for the degrees-of-freedom; rounded up to Inf if above this value}

\item{density.name}{Name of the KDE distribution; used for naming of the probability functions (a character string)}

\item{value.name}{Name of the values in the data; used for naming the plot of the KDE}

\item{discrete}{Logical; if \code{TRUE} the function produces a discrete KDE over the integers}

\item{discrete.warn}{Logical; if \code{TRUE} the function gives a warning if non-discrete data is used to produce a discrete KDE}

\item{to.environment}{Logical; if \code{TRUE} the probability functions are attached to the global environment}

\item{envir}{The environment where the probability functions are loaded (if \code{to.environment} is \code{TRUE})}
}
\value{
A \code{kde} object containing the probability functions for the kernel density estimator
}
\description{
\code{KDE} returns the probability function for the kernel density estimator
}
\details{
The kernel density estimator for a set of input data is obtained by taking a mixture distribution
consisting of a (possibly weighted) combination of kernels.  In this function we compute the KDE
using the kernel of the T-distribution; the function can also estimate a discretised version of the
KDE, taken over the integers, if required.  The degrees-of-freedom and the bandwidth for the KDE can
be specified in the inputs; if they are not specified then they are estimated using leave-one-out
maximum-likelihood estimation (LOO-MLE).  The output of the function is a list of class \code{kde}
that contains the probability functions for the KDE and associated information.  The output object
can be plotted to show the density function for the KDE.

Note: The function has an option \code{to.environment} to allow the user to load the probability
functions to the global environment.  If this is set to \code{TRUE} then the probability functions
are loaded to the global environment in addition to appearing as elements of the output; there is
a message informing the user if existing objects in the global environment were overwritten.

Note: For simplicity, the function uses a value \code{df.norm} which is the maximum value of the
degrees-of-freedom that is recognised as differing from normality.  If the inputted or estimated
degrees-of-freedom is higher than this value then the algorithm will round the value to infinity
(i.e., using the normal-distribution kernel instead of the T-distribution kernel).
}
