% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{ansi_handler}
\alias{ansi_handler}
\alias{create_trimming_hook}
\title{Convert ANSI color codes to HTML tags.}
\usage{
ansi_handler(x, options)

create_trimming_hook(output_hook)
}
\arguments{
\item{x}{Text snippet.}

\item{options}{options.}

\item{output_hook}{\code{knitr}'s output hook used to print the
transformed output.}
}
\description{
This is a callback for \link{knitr}.

Creates an output callback function (hook) for \link{knitr}. When registered,
it gets triggered by the \code{output.lines} option passed in a \code{knitr}'s
code snippet. See this \href{https://stackoverflow.com/questions/23114654/knitr-output-hook-with-an-output-lines-option-that-works-like-echo-26}{Stack Overflow}
question for more details.
}
\examples{
\dontrun{
knitr::knit_hooks$set(output = ansi_handler)
knitr::knit_hooks$set(message = ansi_handler)
}
\dontrun{
output_hook <- knit_hooks$get("output")
trimming_hook <- create_trimming_hook(output_hook)
knit_hooks$set(output = trimming_hook)
}
}
