% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{lapply_preserve_names}
\alias{lapply_preserve_names}
\title{Lapply but re-attach named list entries}
\usage{
lapply_preserve_names(list, fun)
}
\arguments{
\item{list}{list}

\item{fun}{function}
}
\description{
Lapply but re-attach named list entries
}
