% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-list2char.R
\name{list2char}
\alias{list2char}
\title{list2char}
\usage{
list2char(data, except = NULL, sep = NULL)
}
\arguments{
\item{data}{data.table; dane w postaci data.table}

\item{except}{character vector; nazwy kolumn podawane w wektorze, których nie chcemy zamieniać na character}

\item{sep}{character; separator zamienianych kolumn w data.table}
}
\value{
data.table
}
\description{
Funkcja zamieniająca klasę list na string w kolumnach data.table
}
\examples{
data <- data.table(a = c(1,2,3), b = list(c("a", "b"), "c", c("d", "e")))
data_no_list <- list2char(data)
}
