% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{combine}
\alias{combine}
\alias{join}
\title{Merge lists in order.}
\usage{
combine(...)

join(x, sep)
}
\arguments{
\item{...}{Lists to be merged.}

\item{x}{Container of strings to concatenate.}

\item{sep}{Separator.}
}
\value{
A single, reduced \code{list}.
}
\description{
Merge lists in order.

Concatenate multiple strings.
}
