% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design.R
\name{make_design}
\alias{make_design}
\title{Make design matrix from lme4 formula}
\usage{
make_design(formula, data, random_spline = FALSE, ...)
}
\arguments{
\item{formula}{formula using \code{lme4} style, e.g., \code{~ 1 + (0+SST|group)} for a random-slope model}

\item{data}{data-frame used to parse the formula}

\item{random_spline}{Boolean indicating whether to treat basis-spline coefficients as a grouping factor}

\item{...}{additional arguments passed to \code{model.frame}}
}
\value{
Tagged-list of containing a design-matrix \code{X_ij} and the grouping variable
        \code{Group_j} where \code{Group_j[X]=NA} indicates that column \code{X} of \code{X_ij}
        corresponds to a fixed effect, whereas \code{Group_j[X]} having a positive integer
        indicates that column \code{X} is associated with that group of random effects.
}
\description{
\code{make_design} generates a design-matrix and grouping vector from a formula representing fixed and random effects
}
