% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tmbstan_inits.R
\name{make_tmbstan_inits}
\alias{make_tmbstan_inits}
\title{Make initial values list for tmbstan}
\usage{
make_tmbstan_inits(
  obj,
  cores,
  amount = 1,
  upper = rep(Inf, length(obj$par)),
  lower = rep(-Inf, length(obj$par))
)
}
\arguments{
\item{obj}{Compiled TMB object, either before or after optimization}

\item{cores}{the number of cores to initialize}

\item{amount}{difference plus or minus from \code{obj$env$last.par} used for a uniform jitter}

\item{upper}{optional vector of upper bounds for fixed effects}

\item{lower}{optional vector of lower bounds for fixed effects}
}
\description{
\code{make_tmbstan_inits} makes a list of initial values for use by \code{tmbsta::tmbstan}
}
