% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map-helpers}
\alias{map-helpers}
\alias{not}
\title{Mapping helpers.}
\usage{
not(f)
}
\arguments{
\item{f}{function}
}
\description{
\code{not} creates a function that returns the negated value
of \code{f} called with the same list of arguments.
}
\seealso{
\link[utilities:map]{utilities::map}
}
