% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_lst}
\alias{map_lst}
\alias{map_chr}
\alias{map_dbl}
\alias{map_int}
\alias{map_lgl}
\alias{imap}
\title{Purrr-like API.}
\usage{
map_lst(x, f, ...)

map_chr(x, f, ...)

map_dbl(x, f, ...)

map_int(x, f, ...)

map_lgl(x, f, ...)

imap(lst, f, ...)
}
\description{
Purrr-like API.
}
