% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{with_names}
\alias{with_names}
\alias{all_named}
\alias{is_all_named}
\title{Working with names.}
\usage{
with_names(x, names)

all_named(x)

is_all_named(x)
}
\arguments{
\item{x}{Object to check or set \code{names} on.}

\item{names}{Names to set on \code{x}.}
}
\value{
\code{x} with \code{names} set as names.
}
\description{
Working with names.
}
