% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{performGOTermAnalysis}
\alias{performGOTermAnalysis}
\title{Calls HyperGTest}
\usage{
performGOTermAnalysis(bg.genes, geneL, conditional = FALSE,
  pvalueCutoff = 0.1, pValCorMethod = "hochberg", ontology = "BP",
  annotation = "org.Mm.eg.db", testDirection = "over", ...)
}
\arguments{
\item{bg.genes}{background gene list (ENTREZ IDs)}

\item{geneL}{genes to test (May be significant genes from Analysis) (ENTREZ IDs)}

\item{conditional}{(see GOHyperGParams)}

\item{ontology}{(see GOHyperGParams)}

\item{annotation}{(see GOHyperGParams)}

\item{testDirection}{(see GOHyperGParams)}

\item{...}{(see GOHyperGParams)}
}
\value{
List containing the summarized result and the result from the hyperGTest
}
\description{
This function is a wrapper for GO-Term HyperG test. It will allow to modify all parameters for creation of GOHyperGParams object
}

