% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{plot.data.mappings}
\alias{plot.data.mappings}
\title{Plot components from data mapping}
\usage{
\method{plot}{data.mappings}(x, node.size = 1, text.size = 1, line.width = 1, ...)
}
\arguments{
\item{x}{a data mapping}

\item{node.size}{node size}

\item{text.size}{label size for a node}

\item{line.width}{line width}

\item{...}{not used}
}
\value{
nothing
}
\description{
This needs \code{ggplot2} and \code{ggdag} to function correctly.
}
