% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{has_name}
\alias{has_name}
\alias{is_empty}
\alias{is_error}
\alias{is_atomic_class}
\alias{is_recorded_plot}
\title{Accessing properties.}
\usage{
has_name(x, name)

is_empty(x)

is_error(x)

is_atomic_class(x)

is_recorded_plot(x)
}
\description{
Accessing properties.
}
