% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_workbook_to_list.R
\name{read.all.xlsx}
\alias{read.all.xlsx}
\title{Reads an xlsx file into a list of data.frames, one for each sheet.}
\usage{
read.all.xlsx(file = NULL)
}
\arguments{
\item{file}{path to xlsx workbook}
}
\value{
list of data.frame
}
\description{
Reads an xlsx file into a list of data.frames, one for each sheet.
}
