% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_latest.R
\name{read.latest}
\alias{read.latest}
\title{Reads the latest file in a folder.}
\usage{
read.latest(folder = NULL, filetype = "json", pattern = NULL, ...)
}
\arguments{
\item{folder}{path to folder}

\item{filetype}{csv, csv2, json or excel. Defaults to json.}

\item{pattern}{pattern to match files on. Note that by default files are filtered by file extensions correct for the type.}

\item{...}{further arguments passed to jsonlite::fromJSON, read.csv, openxlsx::read.xlsx, etc.}
}
\description{
Reads the latest file in a folder.
}
