% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_workbook_to_list.R
\name{read.xlsx.all}
\alias{read.xlsx.all}
\title{Reads an xlsx file into a list of data.frames, one for each sheet.}
\usage{
read.xlsx.all(file = NULL)
}
\value{
list of data.frame
}
\description{
Reads an xlsx file into a list of data.frames, one for each sheet.
}

