% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-remove_whitespaces.R
\name{remove_whitespaces}
\alias{remove_whitespaces}
\title{remove_whitespaces}
\usage{
remove_whitespaces(string, beginning = FALSE, end = FALSE)
}
\arguments{
\item{string}{string;}

\item{beginning}{logical; if TRUE, then removes whitespaces from the beginnig of string}

\item{end}{logical; jeśli if FALSE, then removes whitespaces from the end of string}
}
\value{
string bez bialych znaków
}
\description{
removes whitespaves such as space, tab, nee line, non-breaking space; is vectorized; if beginnig and end are TRUE, then removes ALL whitespaces from string
}
\examples{
remove_whitespaces("  ad  asd as  ", beginning = FALSE, end = TRUE)
}
