% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns}
\usage{
rename_columns(DF, origname = colnames(DF), newname)
}
\arguments{
\item{DF}{original data frame}

\item{newname}{new names for columns}
}
\description{
\code{rename_columns} renames columns of a matrix or data frame
}
