% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_install.R
\name{Require}
\alias{Require}
\title{Installs a required package if not available.}
\usage{
Require(..., github = NULL)
}
\arguments{
\item{...}{string package name.}

\item{github}{string representation of github archive to install the package from.}
}
\value{
none
}
\description{
Installs a required package if not available.
}

