% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save_fig.R
\name{save_fig}
\alias{save_fig}
\title{Save a figure}
\usage{
save_fig(filename, width, height, units = "in", res = 200, type = "png",
  suffix = "", FUN = NULL, ...)
}
\arguments{
\item{filename}{name (including absolute path) for figure}

\item{width}{width for figure}

\item{height}{height for figure}

\item{units}{units for measuring height/width}

\item{type}{character-vector specifying type of figure (Options: "png", "tif", "jpg"). If \code{length(type)}>1, and \code{FUN} is specified, then makes muliple plots}

\item{suffix}{character-vector specifying suffix for different plots, e.g., if \code{type}>2}

\item{FUN}{optional function containing expression (all lines of script) for plotting.  Useful if \code{length(type)}>1.}
}
\description{
\code{save_fig} saves a figure to a local directory
}

