% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_showtext.R
\name{setup_showtext}
\alias{setup_showtext}
\title{Configure {showtext} options}
\usage{
setup_showtext()
}
\value{
Invisible
}
\description{
A wrapper around several showtext functions to import Google Noto fonts
}
\examples{
setup_showtext()

}
