% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21-source_dir.R
\name{source_dir}
\alias{source_dir}
\title{source_dir}
\usage{
source_dir(path)
}
\arguments{
\item{path}{character; ścieżka do katalagu}
}
\description{
source wszystkich skryptów (*.R) ze wskazanego katalogu
}
