% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_custom}
\alias{theme_custom}
\title{A custom ggplot2 theme}
\usage{
theme_custom(...)
}
\value{
A ggplot2 theme object.
}
\description{
A custom ggplot2 theme built to conform to my (Josh Persi's) personal default tastes.
}
\examples{
temp_file <- fs::file_temp(ext = ".png")

grDevices::png(temp_file, width = 6, height = 6, units = "in", res = 300)

p <- ggplot2::ggplot(penguins, ggplot2::aes(bill_dep, bill_len)) +
  ggplot2::geom_point() +
  theme_custom()

base::print(p)

grDevices::dev.off()

}
