% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-unifyCaseType.R
\name{unify_case_type}
\alias{unify_case_type}
\title{unifyCaseType}
\usage{
unify_case_type(DT_frame, cols2lower = c("ac", "cid", "products"),
  cols2upper = c("country"))
}
\arguments{
\item{DT_frame}{data.table}

\item{cols2lower}{character; wektor nazw kolumn, które mają byc zamienione na LOWER CASE}

\item{cols2upper}{character; wektor nazw kolumn, które mają byc zamienione na UPPER CASE}
}
\value{
data.table
}
\description{
Funkcja w wybrancych kolumanch zmienia wartości na UPPERCASE/LOWERCASE
}
\examples{
data <- data.table(a = c("X", "y", "Z"), b = c("a", "a", "Aa"), c = c("q", "w", "e"))
data2 <- unify_case_type(data, cols2lower = c("a", "b"), cols2upper = "c")
}
