% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cat0}
\alias{cat0}
\alias{ccat}
\alias{ccat0}
\alias{cinform}
\alias{cinform0}
\title{Colored cat().}
\usage{
cat0(..., sep = "", default = "default")

ccat(..., sep = " ", default = "default")

ccat0(..., default = "default")

cinform(..., sep = " ", default = "default")

cinform0(..., default = "default")
}
\arguments{
\item{...}{Strings passed to \code{\link[=cpaste]{cpaste()}}.}

\item{sep}{Separator.}

\item{default}{Default color; \code{"default"} means no special color.}
}
\description{
Colored cat().
}
