% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_names}
\alias{clean_names}
\title{Clean Data Frame Names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{data frame or character vector}
}
\value{
Cleaned data frame or character vector
}
\description{
Applies some general formatting and replacements to enfore a naming
convention that is clear and easy to read, while also ensuring it's valid in
the context of data frame column names.
}
\details{
Essentially functions as a wrapper for make.names, making some useful
substitutions before passing to make.names for final validation.
}
