% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{is_container}
\alias{is_container}
\alias{as_container}
\alias{print.container}
\title{Abstract container.}
\usage{
is_container(x)

as_container(x)

\method{print}{container}(x, ...)
}
\arguments{
\item{x}{object to be converted to a \code{container} or tested.}

\item{...}{Arguments passed on to \code{base::print}
\describe{
  \item{x}{an object used to select a method.}
}}
}
\description{
Used to store artifacts or commits.
}
