% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_to_unicode.R
\name{convert_to_unicode}
\alias{convert_to_unicode}
\title{Convert character to unicode}
\usage{
convert_to_unicode(char, pattern = "-", replacement = "<U+0096>")
}
\arguments{
\item{char}{character-string in which to find a character to replace}

\item{pattern}{character to replace}

\item{replacement}{replacement (presumably using unicode formatting)}
}
\description{
\code{convert_to_unicode} converts a single character to a unicode (e.g., dash to en-hyphen)
}

