% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist_Fn.R, R/zzz.R
\name{Hist_Fn}
\alias{Hist_Fn}
\title{Plot histogram}
\usage{
Hist_Fn(...)

Hist_Fn(...)
}
\arguments{
\item{...}{arguments passed to \code{hist}}

\item{x, }{a list of entries to transform to a histogram and plot}

\item{y_buffer, }{a buffer at the top of the highest histogram}

\item{xlim, }{bounds for x-axis (default is detected from \code{x})}

\item{ylim, }{bounds for y-axis (default is detected from \code{x})}
}
\description{
\code{Hist_Fn} plots one or more histograms on a single panel

Included for continuity when using old scripts
}
\details{
Please use \code{?plot_histogram} to see list of arguments and usage
}
