% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-is_missing.R
\name{is_missing}
\alias{is_missing}
\title{is_missing}
\usage{
is_missing(vec, check_if_finite = TRUE)
}
\arguments{
\item{vec}{vector. Wektor z wartościami, w których spodziewamy się brakujących wartości.}

\item{check_if_finite}{logical. Jeśli flaga jest TRUE, wartość `Inf` też będzie uznawana za brakującą.}
}
\value{
logical vector
}
\description{
Sprawdzenie, czy elementy wektora nie są puste (NA, NULL, NaN, "" lub " "; opcjonalnie: +/-Inf)
}
\examples{
is_missing(vec = c(1,3, NA, NaN, Inf), check_if_finite = FALSE)
}
