% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-list2char.R
\name{list2char}
\alias{list2char}
\title{list2char}
\usage{
list2char(data, except = NULL, sep = NULL)
}
\arguments{
\item{data}{data.table;}

\item{except}{character vector; column names we don't want to change to character}

\item{sep}{character; to separate values; if NULL then value is set to ", ";}
}
\value{
data.table
}
\description{
Changes 'list' class to 'character' in data.table column
}
\examples{
data <- data.table(a = c(1,2,3), b = list(c("a", "b"), "c", c("d", "e")))
data_no_list <- list2char(data)
}
