% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nord_palette.R
\docType{data}
\name{nord_palette}
\alias{nord_palette}
\title{Nord colour palette}
\format{
A named character vector with 16 colors
}
\usage{
nord_palette
}
\description{
A named vector of hex colors
}
\keyword{datasets}
