% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{has_name}
\alias{has_name}
\alias{is_empty}
\alias{is_error}
\alias{is_atomic_class}
\alias{is_recorded_plot}
\alias{is_index_of}
\alias{\%nin\%}
\alias{is_rstudio}
\title{Accessing properties.}
\usage{
has_name(x, name)

is_empty(x)

is_error(x)

is_atomic_class(x)

is_recorded_plot(x)

is_index_of(i, x)

x \%nin\% table

is_rstudio()
}
\arguments{
\item{x}{Object to be checked.}

\item{name}{Name that \code{x} is supposed to have.}

\item{i}{Index \code{x} is supposed to have.}

\item{table}{Container where \code{x} is not supposed to be present.}
}
\description{
\code{is_rstudio} returns \code{TRUE} if is called inside RStudio.
}
