% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-remove_whitespaces.R
\name{remove_whitespaces}
\alias{remove_whitespaces}
\title{remove_whitespaces}
\usage{
remove_whitespaces(string, beginning = FALSE, end = FALSE)
}
\arguments{
\item{string}{string;}

\item{beginning}{logical; jeśli TRUE to usuwa białe znaki z początku stringa}

\item{end}{logical; jeśli TRUE to usuwa białe znaki z końca stringa}
}
\value{
string bez bialych znaków
}
\description{
Funkcja usuwa białe znaki takie jak spacja, tabulator, nowa linia, non-breaking spacja. Funkcja jest zwektoryzowana. Jeśli beginning i end są FALSE to usuwa wszyskie białem znaki ze stringa
}
\examples{
remove_whitespaces("  ad  asd as  ", beginning = FALSE, end = TRUE)
}
