% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_install.R
\name{Require}
\alias{Require}
\title{Installs a required package if not available.}
\usage{
Require(package, ...)
}
\arguments{
\item{package}{string package name}

\item{...}{additional options passed on to \code{\link{install.packages}}.}
}
\value{
none
}
\description{
Installs a required package if not available.
}

