% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_install.R
\name{Require}
\alias{Require}
\title{Installs a required package if not available.}
\usage{
Require(pkg, github = NULL)
}
\arguments{
\item{pkg}{string or literal package name.}

\item{github}{string representation of github archive to install the package from if not found in local library. Will attempt to install from CRAN if left blank.}
}
\value{
none
}
\description{
Installs a required package if not available.
}
