% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{runtime}
\alias{runtime}
\alias{is_knitr}
\alias{try_load}
\alias{is_rstudio}
\title{Runtime checks.}
\usage{
is_knitr()

try_load(package)

is_rstudio()
}
\arguments{
\item{package}{Package name; character or symbol.}
}
\description{
\code{is_knitr} returns \code{TRUE} if invoked while knitr is
compiling a document.

\code{try_load} attempts to load package \code{package} and returns
\code{TRUE} if succeeded; \code{FALSE} otherwise.

\code{is_rstudio} returns \code{TRUE} if is called inside RStudio.
}
\examples{
\dontrun{
  try_load(dplyr)
  try_load("dplyr")
}
}
