% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-unify_case_type.R
\name{unify_case_type}
\alias{unify_case_type}
\title{unify_case_type}
\usage{
unify_case_type(DT_frame, cols2lower, cols2upper)
}
\arguments{
\item{DT_frame}{data.table}

\item{cols2lower}{character vector; column names to convert to LOWERCASE}

\item{cols2upper}{character vector; column names to convert toUPPERCASE}
}
\value{
data.table
}
\description{
converts selected columns to UPPERCASE or LOWERCASE
}
\examples{
data <- data.table(a = c("X", "y", "Z"), b = c("a", "a", "Aa"), c = c("q", "w", "e"))
data2 <- unify_case_type(data, cols2lower = c("a", "b"), cols2upper = "c")
}
